requirements.txt
numpy==2.2.4
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.1.0
python-pptx==1.0.2
scipy==1.15.2
streamlit==1.41.1
matplotlib==3.9.3
google-generativeai
PyMuPDF==1.26.0
anthropic==0.55.0
streamlit-plotly-events==0.0.6